gs: Use common test bench for frontends
