+++ pkg/gatewayserver/io/grpc/grpc_test.go
@@ -17,20 +17,19 @@ package grpc_test
 import (
 	"context"
 	"fmt"
-	"strconv"
 	"sync"
 	"testing"
-	"time"
 
 	"github.com/smarty/assertions"
-	"go.thethings.network/lorawan-stack/v3/pkg/band"
 	"go.thethings.network/lorawan-stack/v3/pkg/cluster"
 	"go.thethings.network/lorawan-stack/v3/pkg/component"
 	componenttest "go.thethings.network/lorawan-stack/v3/pkg/component/test"
 	"go.thethings.network/lorawan-stack/v3/pkg/config"
+	"go.thethings.network/lorawan-stack/v3/pkg/errorcontext"
 	"go.thethings.network/lorawan-stack/v3/pkg/errors"
-	"go.thethings.network/lorawan-stack/v3/pkg/gatewayserver/io"
+	"go.thethings.network/lorawan-stack/v3/pkg/gatewayserver"
 	. "go.thethings.network/lorawan-stack/v3/pkg/gatewayserver/io/grpc"
+	"go.thethings.network/lorawan-stack/v3/pkg/gatewayserver/io/iotest"
 	"go.thethings.network/lorawan-stack/v3/pkg/gatewayserver/io/mock"
 	mockis "go.thethings.network/lorawan-stack/v3/pkg/identityserver/mock"
 	"go.thethings.network/lorawan-stack/v3/pkg/log"
@@ -41,18 +40,15 @@ import (
 	"go.thethings.network/lorawan-stack/v3/pkg/util/test"
 	"go.thethings.network/lorawan-stack/v3/pkg/util/test/assertions/should"
 	"google.golang.org/grpc"
-	"google.golang.org/protobuf/types/known/timestamppb"
-)
-
-var (
-	registeredGatewayID  = &ttnpb.GatewayIdentifiers{GatewayId: "test-gateway"}
-	registeredGatewayUID = unique.ID(test.Context(), registeredGatewayID)
-	registeredGatewayKey = "test-key"
-
-	timeout = (1 << 4) * test.Delay
 )
 
 func TestAuthentication(t *testing.T) {
+	var (
+		registeredGatewayID  = &ttnpb.GatewayIdentifiers{GatewayId: "test-gateway"}
+		registeredGatewayKey = "test-key"
+		timeout              = (1 << 4) * test.Delay
+	)
+
 	ctx := log.NewContext(test.Context(), test.GetLogger(t))
 	ctx, cancelCtx := context.WithCancel(ctx)
 	defer cancelCtx()
@@ -150,465 +146,76 @@ func TestAuthentication(t *testing.T) {
 	}
 }
 
-type erroredGatewayDown struct {
-	*ttnpb.GatewayDown
-	error
-}
-
-func TestTraffic(t *testing.T) {
-	ctx := log.NewContext(test.Context(), test.GetLogger(t))
-	ctx, cancelCtx := context.WithCancel(ctx)
-	defer cancelCtx()
-
-	is, isAddr, closeIS := mockis.New(ctx)
-	defer closeIS()
-	testGtw := mockis.DefaultGateway(registeredGatewayID, false, false)
-	is.GatewayRegistry().Add(ctx, registeredGatewayID, registeredGatewayKey, testGtw, testRights...)
-
-	c := componenttest.NewComponent(t, &component.Config{
-		ServiceBase: config.ServiceBase{
-			GRPC: config.GRPC{
-				Listen:                      ":0",
-				AllowInsecureForCredentials: true,
-			},
-			Cluster: cluster.Config{
-				IdentityServer: isAddr,
-			},
-			FrequencyPlans: config.FrequencyPlansConfig{
-				ConfigSource: "static",
-				Static:       test.StaticFrequencyPlans,
-			},
+func TestFrontend(t *testing.T) {
+	t.Parallel()
+	iotest.Frontend(t, iotest.FrontendConfig{
+		SupportsStatus:      true,
+		IsAuthenticated:     true,
+		DetectsDisconnect:   true,
+		DeduplicatesUplinks: true,
+		Link: func(
+			ctx context.Context,
+			t *testing.T,
+			gs *gatewayserver.GatewayServer,
+			ids *ttnpb.GatewayIdentifiers,
+			key string,
+			upCh <-chan *ttnpb.GatewayUp,
+			downCh chan<- *ttnpb.GatewayDown,
+		) error {
+			md := rpcmetadata.MD{
+				ID:            ids.GatewayId,
+				AuthType:      "Bearer",
+				AuthValue:     key,
+				AllowInsecure: true,
+			}
+			client := ttnpb.NewGtwGsClient(gs.LoopbackConn())
+			_, err := client.GetConcentratorConfig(ctx, ttnpb.Empty, grpc.PerRPCCredentials(md))
+			if err != nil {
+				return err
+			}
+			link, err := client.LinkGateway(ctx, grpc.PerRPCCredentials(md))
+			if err != nil {
+				return err
+			}
+			ctx, cancel := errorcontext.New(ctx)
+			// Write upstream.
+			go func() {
+				for {
+					select {
+					case <-ctx.Done():
+						return
+					case msg := <-upCh:
+						if err := link.Send(msg); err != nil {
+							cancel(err)
+							return
+						}
+					}
+				}
+			}()
+			// Read downstream.
+			go func() {
+				for {
+					msg, err := link.Recv()
+					if err != nil {
+						cancel(err)
+						return
+					}
+					downCh <- msg
+				}
+			}()
+			<-ctx.Done()
+			return ctx.Err()
 		},
 	})
-	gs := mock.NewServer(c, is)
-	srv := New(gs)
-	c.RegisterGRPC(&mockRegisterer{ctx, srv})
-	componenttest.StartComponent(t, c)
-	defer c.Close()
-
-	mustHavePeer(ctx, c, ttnpb.ClusterRole_ENTITY_REGISTRY)
-
-	client := ttnpb.NewGtwGsClient(c.LoopbackConn())
-
-	ctx = rpcmetadata.MD{
-		ID: registeredGatewayID.GatewayId,
-	}.ToOutgoingContext(ctx)
-	creds := grpc.PerRPCCredentials(rpcmetadata.MD{
-		AuthType:      "Bearer",
-		AuthValue:     registeredGatewayKey,
-		AllowInsecure: true,
-	})
-
-	upCh := make(chan *ttnpb.GatewayUp, 10)
-	downCh := make(chan erroredGatewayDown, 10)
-
-	stream, err := client.LinkGateway(ctx, creds)
-	if err != nil {
-		t.Fatalf("Failed to link gateway: %v", err)
-	}
-	go func() {
-		for up := range upCh {
-			if err := stream.Send(up); err != nil {
-				panic(err)
-			}
-		}
-	}()
-	go func() {
-		for ctx.Err() == nil {
-			down, err := stream.Recv()
-			downCh <- erroredGatewayDown{down, err}
-		}
-	}()
-
-	var conn *io.Connection
-	select {
-	case conn = <-gs.Connections():
-	case <-time.After(timeout):
-		t.Fatal("Connection timeout")
-	}
-
-	t.Run("Upstream", func(t *testing.T) {
-		for _, tc := range []*ttnpb.GatewayUp{
-			{},
-			{
-				GatewayStatus: &ttnpb.GatewayStatus{
-					Ip:   []string{"1.1.1.1"},
-					Time: timestamppb.Now(),
-				},
-			},
-			{
-				UplinkMessages: []*ttnpb.UplinkMessage{
-					{
-						RawPayload: []byte{0x01},
-						RxMetadata: []*ttnpb.RxMetadata{
-							{
-								GatewayIds: registeredGatewayID,
-							},
-						},
-						Settings: &ttnpb.TxSettings{
-							DataRate: &ttnpb.DataRate{
-								Modulation: &ttnpb.DataRate_Lora{Lora: &ttnpb.LoRaDataRate{
-									Bandwidth:       125000,
-									SpreadingFactor: 11,
-									CodingRate:      band.Cr4_5,
-								}},
-							},
-							EnableCrc: true,
-							Frequency: 868500000,
-							Timestamp: 42,
-						},
-					},
-				},
-			},
-			{
-				UplinkMessages: []*ttnpb.UplinkMessage{
-					{
-						RawPayload: []byte{0x02},
-						RxMetadata: []*ttnpb.RxMetadata{
-							{
-								GatewayIds: registeredGatewayID,
-							},
-						},
-						Settings: &ttnpb.TxSettings{
-							DataRate: &ttnpb.DataRate{
-								Modulation: &ttnpb.DataRate_Lora{Lora: &ttnpb.LoRaDataRate{
-									Bandwidth:       125000,
-									SpreadingFactor: 11,
-									CodingRate:      band.Cr4_5,
-								}},
-							},
-							EnableCrc: true,
-							Frequency: 868500000,
-							Timestamp: 42,
-						},
-					},
-				},
-				GatewayStatus: &ttnpb.GatewayStatus{
-					Ip:   []string{"2.2.2.2"},
-					Time: timestamppb.Now(),
-				},
-			},
-			{
-				UplinkMessages: []*ttnpb.UplinkMessage{
-					{
-						RawPayload: []byte{0x03},
-						RxMetadata: []*ttnpb.RxMetadata{
-							{
-								GatewayIds: registeredGatewayID,
-							},
-						},
-						Settings: &ttnpb.TxSettings{
-							DataRate: &ttnpb.DataRate{
-								Modulation: &ttnpb.DataRate_Lora{Lora: &ttnpb.LoRaDataRate{
-									Bandwidth:       125000,
-									SpreadingFactor: 11,
-									CodingRate:      band.Cr4_5,
-								}},
-							},
-							EnableCrc: true,
-							Frequency: 868500000,
-							Timestamp: 42,
-						},
-					},
-					{
-						RawPayload: []byte{0x04},
-						RxMetadata: []*ttnpb.RxMetadata{
-							{
-								GatewayIds: registeredGatewayID,
-							},
-						},
-						Settings: &ttnpb.TxSettings{
-							DataRate: &ttnpb.DataRate{
-								Modulation: &ttnpb.DataRate_Lora{Lora: &ttnpb.LoRaDataRate{
-									Bandwidth:       125000,
-									SpreadingFactor: 11,
-									CodingRate:      band.Cr4_5,
-								}},
-							},
-							EnableCrc: true,
-							Frequency: 868500000,
-							Timestamp: 42,
-						},
-					},
-					{
-						RawPayload: []byte{0x05},
-						RxMetadata: []*ttnpb.RxMetadata{
-							{
-								GatewayIds: registeredGatewayID,
-							},
-						},
-						Settings: &ttnpb.TxSettings{
-							DataRate: &ttnpb.DataRate{
-								Modulation: &ttnpb.DataRate_Lora{Lora: &ttnpb.LoRaDataRate{
-									Bandwidth:       125000,
-									SpreadingFactor: 11,
-									CodingRate:      band.Cr4_5,
-								}},
-							},
-							EnableCrc: true,
-							Frequency: 868500000,
-							Timestamp: 42,
-						},
-					},
-				},
-				GatewayStatus: &ttnpb.GatewayStatus{
-					Ip:   []string{"3.3.3.3"},
-					Time: timestamppb.Now(),
-				},
-			},
-			{
-				TxAcknowledgment: &ttnpb.TxAcknowledgment{
-					Result: ttnpb.TxAcknowledgment_SUCCESS,
-				},
-			},
-		} {
-			t.Run(fmt.Sprintf("%v/%v", len(tc.UplinkMessages), tc.GatewayStatus != nil), func(t *testing.T) {
-				a := assertions.New(t)
-
-				upCh <- tc
-
-				var ups int
-				needStatus := tc.GatewayStatus != nil
-				needTxAck := tc.TxAcknowledgment != nil
-				for ups != len(tc.UplinkMessages) || needStatus || needTxAck {
-					select {
-					case up := <-conn.Up():
-						expected := ttnpb.Clone(tc.UplinkMessages[ups])
-						expected.ReceivedAt = up.Message.ReceivedAt
-						for i, md := range expected.RxMetadata {
-							md.UplinkToken = up.Message.RxMetadata[i].UplinkToken
-							md.ReceivedAt = up.Message.ReceivedAt
-						}
-						a.So(up.Message, should.Resemble, expected)
-						ups++
-					case status := <-conn.Status():
-						a.So(needStatus, should.BeTrue)
-						a.So(status, should.Resemble, tc.GatewayStatus)
-						needStatus = false
-					case ack := <-conn.TxAck():
-						a.So(needTxAck, should.BeTrue)
-						a.So(ack, should.Resemble, tc.TxAcknowledgment)
-						needTxAck = false
-					case <-time.After(timeout):
-						t.Fatalf("Receive expected upstream timeout; ups = %v, needStatus = %t, needAck = %t", ups, needStatus, needTxAck)
-					}
-				}
-			})
-		}
-		t.Run("DeduplicateByRSSI", func(t *testing.T) {
-			a := assertions.New(t)
-			upCh <- &ttnpb.GatewayUp{
-				UplinkMessages: []*ttnpb.UplinkMessage{
-					{
-						RawPayload: []byte{0x06},
-						RxMetadata: []*ttnpb.RxMetadata{{GatewayIds: registeredGatewayID, Rssi: -100}},
-						Settings: &ttnpb.TxSettings{
-							DataRate: &ttnpb.DataRate{
-								Modulation: &ttnpb.DataRate_Lora{Lora: &ttnpb.LoRaDataRate{
-									Bandwidth:       125000,
-									SpreadingFactor: 11,
-									CodingRate:      band.Cr4_5,
-								}},
-							},
-							EnableCrc: true,
-							Frequency: 868500000,
-							Timestamp: 42,
-						},
-					},
-					{
-						RawPayload: []byte{0x06},
-						RxMetadata: []*ttnpb.RxMetadata{{GatewayIds: registeredGatewayID, Rssi: -10}},
-						Settings: &ttnpb.TxSettings{
-							DataRate: &ttnpb.DataRate{
-								Modulation: &ttnpb.DataRate_Lora{Lora: &ttnpb.LoRaDataRate{
-									Bandwidth:       125000,
-									SpreadingFactor: 11,
-									CodingRate:      band.Cr4_5,
-								}},
-							},
-							EnableCrc: true,
-							Frequency: 868700000,
-							Timestamp: 42,
-						},
-					},
-				},
-			}
-			select {
-			case up := <-conn.Up():
-				a.So(up.Message.RxMetadata[0].Rssi, should.Equal, -10)
-				a.So(up.Message.RawPayload, should.Resemble, []byte{0x06})
-				a.So(up.Message.Settings.Frequency, should.Equal, 868700000)
-			case <-time.After(timeout):
-				t.Fatalf("Receive unexpected upstream timeout")
-			}
-			select {
-			case <-conn.Up():
-				t.Fatalf("Received unexpected upstream message")
-			case <-time.After(timeout):
-			}
-		})
-	})
-
-	t.Run("Downstream", func(t *testing.T) {
-		for i, tc := range []struct {
-			Path                *ttnpb.DownlinkPath
-			Message             *ttnpb.DownlinkMessage
-			ErrorAssertion      func(error) bool
-			SendTxAck           bool
-			TxAckErrorAssertion func(error) bool
-		}{
-			{
-				Path: &ttnpb.DownlinkPath{
-					Path: &ttnpb.DownlinkPath_UplinkToken{
-						UplinkToken: io.MustUplinkToken(
-							&ttnpb.GatewayAntennaIdentifiers{GatewayIds: registeredGatewayID},
-							100,
-							100000,
-							time.Unix(0, 100*1000),
-							nil,
-						),
-					},
-				},
-				Message: &ttnpb.DownlinkMessage{
-					RawPayload: []byte{0x01},
-					Settings: &ttnpb.DownlinkMessage_Request{
-						Request: &ttnpb.TxRequest{
-							Class:    ttnpb.Class_CLASS_A,
-							Priority: ttnpb.TxSchedulePriority_NORMAL,
-							Rx1Delay: ttnpb.RxDelay_RX_DELAY_1,
-							Rx1DataRate: &ttnpb.DataRate{
-								Modulation: &ttnpb.DataRate_Lora{
-									Lora: &ttnpb.LoRaDataRate{
-										SpreadingFactor: 7,
-										Bandwidth:       125000,
-										CodingRate:      band.Cr4_5,
-									},
-								},
-							},
-							Rx1Frequency: 868100000,
-							Rx2DataRate: &ttnpb.DataRate{
-								Modulation: &ttnpb.DataRate_Lora{
-									Lora: &ttnpb.LoRaDataRate{
-										SpreadingFactor: 0,
-										Bandwidth:       125000,
-										CodingRate:      band.Cr4_5,
-									},
-								},
-							},
-							Rx2Frequency:    869525000,
-							FrequencyPlanId: test.EUFrequencyPlanID,
-						},
-					},
-				},
-				SendTxAck: true,
-			},
-			{
-				Path: &ttnpb.DownlinkPath{
-					Path: &ttnpb.DownlinkPath_UplinkToken{
-						UplinkToken: io.MustUplinkToken(
-							&ttnpb.GatewayAntennaIdentifiers{GatewayIds: registeredGatewayID},
-							100,
-							100000,
-							time.Unix(0, 100*1000),
-							nil,
-						),
-					},
-				},
-				Message: &ttnpb.DownlinkMessage{
-					RawPayload: []byte{0x01},
-					Settings: &ttnpb.DownlinkMessage_Scheduled{
-						Scheduled: &ttnpb.TxSettings{
-							DataRate: &ttnpb.DataRate{
-								Modulation: &ttnpb.DataRate_Lora{
-									Lora: &ttnpb.LoRaDataRate{
-										Bandwidth:       125000,
-										SpreadingFactor: 7,
-										CodingRate:      band.Cr4_5,
-									},
-								},
-							},
-							Frequency: 869525000,
-						},
-					},
-				},
-				ErrorAssertion: errors.IsInvalidArgument, // Does not support scheduled downlink; the Gateway Server or gateway will take care of that.
-			},
-			{
-				Path: &ttnpb.DownlinkPath{
-					Path: &ttnpb.DownlinkPath_UplinkToken{
-						UplinkToken: io.MustUplinkToken(
-							&ttnpb.GatewayAntennaIdentifiers{GatewayIds: registeredGatewayID},
-							100,
-							100000,
-							time.Unix(0, 100*1000),
-							nil,
-						),
-					},
-				},
-				Message: &ttnpb.DownlinkMessage{
-					RawPayload: []byte{0x02},
-				},
-				ErrorAssertion: errors.IsInvalidArgument, // Tx request missing.
-			},
-		} {
-			t.Run(strconv.Itoa(i), func(t *testing.T) {
-				a := assertions.New(t)
-
-				_, _, _, err := conn.ScheduleDown(tc.Path, tc.Message)
-				if err != nil && (tc.ErrorAssertion == nil || !a.So(tc.ErrorAssertion(err), should.BeTrue)) {
-					t.Fatalf("Unexpected error: %v", err)
-				}
-				var cids []string
-				select {
-				case down := <-downCh:
-					if tc.ErrorAssertion == nil {
-						cids = down.DownlinkMessage.CorrelationIds
-						a.So(down.DownlinkMessage, should.Resemble, tc.Message)
-					} else {
-						t.Fatalf("Unexpected message: %v", down.DownlinkMessage)
-					}
-				case <-time.After(timeout):
-					if tc.ErrorAssertion == nil {
-						t.Fatal("Receive expected downlink timeout")
-					}
-				}
-
-				if tc.ErrorAssertion != nil || !tc.SendTxAck {
-					return
-				}
-				select {
-				case upCh <- &ttnpb.GatewayUp{
-					TxAcknowledgment: &ttnpb.TxAcknowledgment{
-						CorrelationIds: cids,
-						Result:         ttnpb.TxAcknowledgment_SUCCESS,
-					},
-				}:
-				case <-time.After(timeout):
-					if tc.TxAckErrorAssertion == nil {
-						t.Fatal("Receive unexpected timeout while sending Tx acknowledgment")
-					}
-				}
-
-				select {
-				case ack := <-conn.TxAck():
-					a.So(ack.DownlinkMessage, should.Resemble, tc.Message)
-					a.So(ack.Result, should.Equal, ttnpb.TxAcknowledgment_SUCCESS)
-				case <-time.After(timeout):
-					if tc.TxAckErrorAssertion == nil {
-						t.Fatal("Timeout waiting for Tx acknowledgment")
-					}
-				}
-			})
-		}
-	})
-
-	cancelCtx()
 }
 
 func TestConcentratorConfig(t *testing.T) {
-	a := assertions.New(t)
+	var (
+		registeredGatewayID  = &ttnpb.GatewayIdentifiers{GatewayId: "test-gateway"}
+		registeredGatewayKey = "test-key"
+	)
 
-	ctx := log.NewContext(test.Context(), test.GetLogger(t))
+	a, ctx := test.New(t)
 	ctx, cancelCtx := context.WithCancel(ctx)
 	defer cancelCtx()
 
@@ -664,9 +271,13 @@ func (p mockMQTTConfigProvider) GetMQTTConfig(context.Context) (*config.MQTT, er
 }
 
 func TestMQTTConfig(t *testing.T) {
-	a := assertions.New(t)
+	var (
+		registeredGatewayID  = &ttnpb.GatewayIdentifiers{GatewayId: "test-gateway"}
+		registeredGatewayUID = unique.ID(test.Context(), registeredGatewayID)
+		registeredGatewayKey = "test-key"
+	)
 
-	ctx := log.NewContext(test.Context(), test.GetLogger(t))
+	a, ctx := test.New(t)
 	ctx, cancelCtx := context.WithCancel(ctx)
 	defer cancelCtx()
